server/requirements.txt
Flask
flask_prometheus_metrics
pandas
prometheus_client
pymongo
pytest
swagger-ui-py
Werkzeug
